+++ src/0_maze_gen/goal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   goal.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
+/*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/28 20:25:23 by freddy            #+#    #+#             */
-/*   Updated: 2024/07/01 11:45:14 by freddy           ###   ########.fr       */
+/*   Updated: 2024/07/02 11:40:11 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,9 +36,10 @@ static inline void	place_door_and_goal(char **maze, t_scale door1, \
 {
 	char	door_dir;
 
-	door_dir = 'H';
+	printf("placing gooal wooho\n");
+	door_dir = 'V';
 	if (door1.x == door2.x)
-		door_dir = 'V';
+		door_dir = 'H';
 	maze[door1.y][door1.x] = door_dir;
 	maze[door2.y][door2.x] = door_dir;
 	maze[goal.y][goal.x] = 'G';